README: [small fix] of `HexagonOffsetGrid.oddPointy` attribute name
There is no `buildHexagon` any more so I renamed it to valid `buildTile` attribute. Although there is `hexagonBuilder`, but it is mentioned in the text bellow code snippet, so I think `buildTile` is the intended one for the documentation code snippet, as there is an intent to use `(col, row)` parameters.